changed data types from int to long ints

+++ 0x14-bit_manipulation/2-get_bit.c
@@ -7,7 +7,7 @@
  *
  * Return: the string or NULL if malloc failed
 */
-char *to_binary(int n)
+char *to_binary(unsigned long int n)
 {
 	int length = 0, n_copy = n;
 	char *binary;
